start moving to activations.py
